feat: 스토리북 목록 카드 배지에 따른 화면 이동 연결

+++ app/src/main/java/com/umc/halo/domain/model/storybook/Storybook.kt
@@ -6,7 +6,7 @@ package com.umc.halo.domain.model.storybook
  * 온보딩에서 고른 관계 방향 태그를 기반으로 추천되며 전체 탭에서만 2개 노출
  */
 data class CustomStorybook(
-    val id: Int,
+    val id: Long,
     val tag: String,       // 추천 이유 태그 (예: "대화가 어색한 당신을 위한")
     val title: String,     // 스토리북 제목 (예: "오래전 당신")
     val subtitle: String   // 부제 (예: "가족과의 만남")
@@ -28,7 +28,7 @@ sealed interface StorybookProgress {
  * 테마 섹션 안에 들어가는 스토리북
  */
 data class Storybook(
-    val id: Int,
+    val id: Long,
     val title: String,
     val subtitle: String,
     val progress: StorybookProgress? = null  // 전체 탭 카드 배지용 (null = 배지 없음)
@@ -39,7 +39,7 @@ data class Storybook(
  * "대화가 어색할 때" 처럼 상황별로 스토리북을 묶은 테마 섹션
  */
 data class StorybookTheme(
-    val id: Int,
+    val id: Long,
     val title: String,               // 섹션 제목 (예: "대화가 어색할 때")
     val storybooks: List<Storybook>  // 해당 상황에 추천되는 책들
 )
@@ -49,7 +49,7 @@ data class StorybookTheme(
  * 목록은 시작일이 빠른 순으로 좌상단부터 배치 (정렬은 데이터 계층에서)
  */
 data class InProgressStorybook(
-    val id: Int,
+    val id: Long,
     val title: String,
     val subtitle: String,
     val currentChapter: Int,  // "N장 진행중" 책갈피에 표시
